Separate tests + key checkpoints by AuthorityName
